agents/chat/requirements.txt
google-adk>=1.10.0
a2a-sdk>=0.2.5
google-cloud-bigquery
google-cloud-logging
python-dotenv
requests
uvicorn
fastapi
nest-asyncio>=1.6.0
httpx
google-generativeai
